requirements.txt
python-twitter
pytest
tweepy
Flask
Flask-SQLAlchemy
SQLAlchemy-serializer